Move CLC test to test_core
